added logparams to job logs, added bool to zip logs
