+++ src/robusta/core/model/base_params.py
@@ -255,3 +255,22 @@ class ProcessParams(ActionParams):
 class EventEnricherParams(ActionParams):
     max_events: int = 8
     included_types: List[str] = ["Warning", "Normal"]
+
+
+class LogEnricherParams(ActionParams):
+    """
+    :var container_name: Specific container to get logs from
+    :var warn_on_missing_label: Send a warning if the alert doesn't have a pod label
+    :var regex_replacer_patterns: regex patterns to replace text, for example for security reasons (Note: Replacements are executed in the given order)
+    :var regex_replacement_style: one of SAME_LENGTH_ASTERISKS or NAMED (See RegexReplacementStyle)
+    :var filter_regex: only shows lines that match the regex
+    :var compress_logs: compresses the logs if true, this is currently only implemented for robusta sinks
+    """
+
+    container_name: Optional[str]
+    warn_on_missing_label: bool = False
+    regex_replacer_patterns: Optional[List[NamedRegexPattern]] = None
+    regex_replacement_style: Optional[str] = None
+    previous: bool = False
+    filter_regex: Optional[str] = None
+    compress_logs: bool = False
